pyproject.toml
[project]
name = "spot-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask",
    "flask-socketio",
    "gevent-websocket",
    "paho-mqtt",
    "ultralytics",
    "opencv-python",
    "Pillow",
]
